Move budget table controller hooks
